+++ bench/pubsub_fanout/bench.rb
@@ -21,7 +21,7 @@ SUBSCRIBER_COUNTS.each do |n_subs|
     OMQ::Transport::Inproc.reset!
 
     pub = OMQ::PUB.bind("inproc://bench_fanout_#{n_subs}")
-    subs = n_subs.times.map { OMQ::SUB.connect("inproc://bench_fanout_#{n_subs}", prefix: "") }
+    subs = n_subs.times.map { OMQ::SUB.connect("inproc://bench_fanout_#{n_subs}", subscribe: "") }
 
     sleep 0.01